Working on Display Ranges
Created new SQL Tables, may still need to work on these.  Figured out
how to manipulate the form using the plaintext option.  Working on
adjusting how the logic is going to work for callbacks and for adding
input to the other tables.

+++ sqlFiles/migrations/2015.04.07.sql
@@ -22,7 +22,7 @@ CREATE TABLE `displayConditions` (
   `months` varchar(50) DEFAULT NULL,
   `day` tinyint(2) DEFAULT NULL,
   `timeStart` tinyint(4) DEFAULT NULL,
-  `year` tinyint(4) DEFAULT NULL
+  `year` int(4) DEFAULT NULL,
   PRIMARY KEY (`ID`)
 ) ENGINE=InnoDB AUTO_INCREMENT=3 DEFAULT CHARSET=latin1;
